separate out backend setup/info
