+++ app/src/main/java/mozilla/lockbox/store/DataStore.kt
@@ -10,11 +10,10 @@ import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.addTo
 import io.reactivex.subjects.ReplaySubject
 import io.reactivex.subjects.SingleSubject
-import java.util.Date
-import java.util.Random
 import mozilla.lockbox.flux.Dispatcher
+import mozilla.lockbox.support.DataStoreSupport
+import mozilla.lockbox.support.FixedDataStoreSupport
 import org.mozilla.sync15.logins.LoginsStorage
-import org.mozilla.sync15.logins.MemoryLoginsStorage
 import org.mozilla.sync15.logins.ServerPassword
 import org.mozilla.sync15.logins.SyncResult
 
@@ -33,7 +32,10 @@ data class DataStoreState(
 /**
  *
  */
-class DataStore(val dispatcher: Dispatcher = Dispatcher.shared) {
+class DataStore(
+        val dispatcher: Dispatcher = Dispatcher.shared,
+        val support: DataStoreSupport = FixedDataStoreSupport.shared
+) {
     internal val compositeDisposable = CompositeDisposable()
     private val stateSubject: ReplaySubject<DataStoreState> = ReplaySubject.create(1)
     private val listSubject: BehaviorRelay<List<ServerPassword>> = BehaviorRelay.createDefault(emptyList())
@@ -42,7 +44,7 @@ class DataStore(val dispatcher: Dispatcher = Dispatcher.shared) {
 
     init {
         // TODO: Replace test data with real backend
-        backend = setupBackend()
+        backend = support.createLoginsStorage()
 
         // handle state changes
         state.subscribe { state ->
@@ -63,7 +65,7 @@ class DataStore(val dispatcher: Dispatcher = Dispatcher.shared) {
 
         backend.isLocked().then {
             if (it) {
-                backend.unlock(getEncryptionKey()).then {
+                backend.unlock(support.encryptionKey).then {
                     stateSubject.onNext(DataStoreState(DataStoreState.Status.UNLOCKED))
                     SyncResult.fromValue(Unit)
                 }
@@ -104,24 +106,6 @@ class DataStore(val dispatcher: Dispatcher = Dispatcher.shared) {
         return lockSubject.toObservable()
     }
 
-    // backend management
-    private fun setupBackend(): LoginsStorage {
-        val testdata = List<ServerPassword>(10) { createTestItem(it) }
-        return MemoryLoginsStorage(testdata)
-    }
-    private fun loadInfo() {
-        // TODO: load info from SharedPreferences ...
-
-        backend.isLocked().whenComplete { result ->
-            val status = if (result) DataStoreState.Status.LOCKED else DataStoreState.Status.UNLOCKED
-            stateSubject.onNext(DataStoreState(status))
-        }
-    }
-    private fun getEncryptionKey(): String {
-        // TODO: read this from SharedPreferences
-        return "keep-this-secret"
-    }
-
     // item list management
     private fun clearList() {
         this.listSubject.accept(emptyList())
@@ -129,32 +113,4 @@ class DataStore(val dispatcher: Dispatcher = Dispatcher.shared) {
     private fun updateList() {
         backend.list().whenComplete { all -> this.listSubject.accept(all) }
     }
-}
-
-internal fun createTestList(amt: Int = 10): List<ServerPassword> {
-    return List(amt) { createTestItem(it) }
-}
-/**
- * Creates a test ServerPassword item
- */
-private fun createTestItem(idx: Int): ServerPassword {
-    val rng = Random()
-    val pos = idx + 1
-    val pwd = "AAbbcc112233!"
-    val host = "https://$pos.example.com"
-    val created = Date().time
-    val used = Date(created - 86400000).time
-    val changed = Date(used - 86400000).time
-
-    return ServerPassword(
-            id = "0000$idx",
-            hostname = host,
-            username = "someone #$pos",
-            password = pwd,
-            formSubmitURL = host,
-            timeCreated = created,
-            timeLastUsed = used,
-            timePasswordChanged = changed,
-            timesUsed = rng.nextInt(100) + 1
-    )
 }
